諸々修正

+++ case2/30-wrapper.py
@@ -6,19 +6,17 @@ import glob
 import random
 
 import sys
- 
+
+from pathlib import Path
+
 confs = glob.glob('./files/confs/*')
 
 random.shuffle(confs)
 
-conf = confs.pop()
-
-ha = conf.split('/').pop()
-os.system(f'lightgbm config={conf}') 
-sys.exit() 
-
-result = os.popen(f'lightgbm config={conf}').read()
-open(f'./files/results/{ha}', 'w').write( result )
-
-sys.exit() 
+for conf in confs:
+  ha = conf.split('/').pop()
+  if Path(f'./files/results/{ha}').exists():
+    continue
+  result = os.popen(f'lightgbm config={conf}').read()
+  open(f'./files/results/{ha}', 'w').write( result )
